requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pytz==2023.3
google-cloud-bigquery==3.10.0